requirements_idea_agent.txt
langchain
langchain-core
langchain-deepseek
pydantic
python-dotenv
langgraph
langchain-community
wikipedia
tavily-python
typing-extensions
